task/requirement.txt
torch==1.12.1
torchvision==0.13.1
numpy==1.23.0
Pillow==9.0.0
h5py==3.6.0
matplotlib==3.5.1
scikit-learn==1.1.1
opencv-python==4.5.5.64
pandas==1.4.2
seaborn==0.11.2
tqdm==4.64.0
